Make small fix to inodes to hopefully fix error
Untested change, but it makes sense. The issue was that you
can't just create new inode blocks and use block[0] each time
obviously, since after the first one the newinodenum found in the
bitmap will be a component of a block. Have to use mod function
to find the right index, AND call Disk_Read after creating the
Inode* block. Adding that every time should fix it I hope...